+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## Unreleased
 
+## 1.11.19
+
+- Minor: Include Soup pet name in pet notifications. (#844)
+- Bugfix: Prevent always-dropped items from appearing as rare due to RDT overlap. (#843)
+
+## 1.11.18
+
+- Minor: Add `%COUNT%` template variable for loot notifier to display killcount when rich embeds are disabled. (#840)
+- Minor: Allow custom author badge image in Discord rich embeds. (#834)
+- Minor: Add configurable minimum level to warrant a screenshot for skill notifier. (#833)
+- Minor: Add drop source denylist for loot notifier. (#835)
+- Minor: Add configurable denylist for collection log notifier. (#832)
+
 ## 1.11.17
 
 - Minor: Unhide the leagues webhook URL for seasonal usage. (#825)